move the util functions to its own module
